Convert multitype parameters to strings
Some parameters could be set to false or a string value. We prefer
consistent type so all such parameters are now forced to be string. In
case you want to indicate that the parameter should not be used you can
set it to undef.

Some documentation was added.

Change lookup to instance variables.